split controller and objects
