+++ src/sc/src/controllers/spus/controller.rs
@@ -37,7 +37,7 @@ impl SpuOnlineStatus {
     }
 }
 
-/// Keep track of SPU status
+/// Keep track of SPU status.  This only updates SPU status.
 /// if SPU has not send heart beat within a period, it is considered down
 pub struct SpuController {
     spus: StoreContext<SpuSpec>,
